Added filebacked and fork test cases. Fixed bug with copy on write block numbers.

+++ vm_globals.cc
@@ -21,7 +21,7 @@ vm_globals::~vm_globals(){
 //		   otherwise reserves number of swap_blocks parent has.
 //Returns: true if enough swap blocks, else false.
 bool vm_globals::reserve_blocks(pid_t parent){
-	int reserve = 1;
+	size_t reserve = 1;
 	if(parent)
 		reserve = app_map[parent]->swap_blocks_used;
 
@@ -60,7 +60,6 @@ bool vm_globals::load_page(unsigned int vpage, char* buffer){
 		while(evicted == nullptr){
 			evicted = clock.front();
 			clock.pop_front();
-			//cout << "clock sweep: " << evicted->pte.ppage << endl;
 
 			//reset read and write enable to fault on next reference
 			evicted->pte.read_enable = 0;
